+++ zcash_client_sqlite/src/lib.rs
@@ -305,7 +305,29 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters> InputSource for 
         &self,
         outpoint: &OutPoint,
     ) -> Result<Option<WalletTransparentOutput>, Self::Error> {
-        wallet::transparent::get_wallet_transparent_output(self.conn.borrow(), outpoint, false)
+        wallet::transparent::get_wallet_transparent_output(
+            self.conn.borrow(),
+            outpoint,
+            false,
+        )
+    }
+
+    /// Fetches the transparent output corresponding to the provided `outpoint`.
+    /// Allows selecting unspendable outputs for testing purposes.
+    ///
+    /// Returns `Ok(None)` if the UTXO is not known to belong to the wallet or is not
+    /// spendable as of the chain tip height.
+    #[cfg(all(feature = "test-dependencies", feature = "transparent-inputs"))]
+    fn get_all_unspent_transparent_output(
+        &self,
+        outpoint: &OutPoint,
+        allow_unspendable: bool,
+    ) -> Result<Option<WalletTransparentOutput>, Self::Error> {
+        wallet::transparent::get_wallet_transparent_output(
+            self.conn.borrow(),
+            outpoint,
+            allow_unspendable,
+        )
     }
 
     #[cfg(feature = "transparent-inputs")]
@@ -333,7 +355,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters> InputSource for 
         use wallet::common::per_protocol_names;
 
         let (table_prefix, index_col, note_reconstruction_cols) = per_protocol_names(protocol);
-        let mut stmt_sent_notes = self.conn.borrow().prepare(&format!(
+        let mut stmt_received_notes = self.conn.borrow().prepare(&format!(
             "SELECT txid, {index_col}
                  FROM {table_prefix}_received_notes rn
                  INNER JOIN transactions ON transactions.id_tx = rn.tx
@@ -343,7 +365,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters> InputSource for 
                  AND commitment_tree_position IS NOT NULL"
         ))?;
 
-        let result = stmt_sent_notes
+        let result = stmt_received_notes
             .query([])
             .unwrap()
             .mapped(|row| {
@@ -678,6 +700,37 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters> WalletRead for W
             .query(named_params![":txid": txid.as_ref(), ":pool_code": pool_code(PoolType::Shielded(protocol))])
             .unwrap()
             .mapped(|row| Ok(NoteId::new(*txid, protocol, row.get(0)?)))
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(SqliteClientError::from)
+    }
+
+    #[cfg(any(test, feature = "test-dependencies"))]
+    fn get_confirmed_sends(
+        &self,
+        txid: &TxId,
+    ) -> Result<Vec<(u64, Option<String>, Option<String>, Option<u32>)>, Self::Error> {
+        let mut stmt_sent = self
+            .conn.borrow()
+            .prepare(
+                "SELECT value, to_address, ephemeral_addresses.address, ephemeral_addresses.address_index
+                 FROM sent_notes
+                 JOIN transactions ON transactions.id_tx = sent_notes.tx
+                 LEFT JOIN ephemeral_addresses ON ephemeral_addresses.used_in_tx = sent_notes.tx
+                 WHERE transactions.txid = ?
+                 ORDER BY value",
+            )
+            .unwrap();
+
+        stmt_sent
+            .query(rusqlite::params![txid.as_ref()])
+            .unwrap()
+            .mapped(|row| {
+                let v: u32 = row.get(0)?;
+                let to_address: Option<String> = row.get(1)?;
+                let ephemeral_address: Option<String> = row.get(2)?;
+                let address_index: Option<u32> = row.get(3)?;
+                Ok((u64::from(v), to_address, ephemeral_address, address_index))
+            })
             .collect::<Result<Vec<_>, _>>()
             .map_err(SqliteClientError::from)
     }
